+++ plugin/src/Plugins/Extensions/PiconHubWarderEvolution/core/downloader.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Streaming download, integrity and atomic-install helpers."""